refactor(ui): MetricsTable - move job header inside component
