--- packages/ui/src/components/bundle-assets-totals-table/bundle-assets-totals-table.module.css
@@ -1,8 +1,0 @@
-.tableHeaderRun {
-  margin-top: calc(0px - var(--space-xxsmall));
-  margin-bottom: calc(0px - var(--space-xxsmall));
-}
-
-.tableHeaderRunMetric {
-  margin-top: var(--space-xxxsmall);
-}
