+++ packages/ui/src/components/bundle-packages/bundle-packages.jsx
@@ -7,37 +7,16 @@ import { EmptySet } from '../../ui/empty-set';
 import { FiltersDropdown } from '../../ui/filters-dropdown';
 import { SortDropdown } from '../../ui/sort-dropdown';
 import { MetricsTable } from '../metrics-table';
-import { JobName } from '../job-name';
-import { RunLabelSum } from '../run-label-sum';
 import { FILTER_CHANGED, FILTER_DUPLICATE } from './bundle-packages.constants';
 import css from './bundle-packages.module.css';
 
-const RUN_TITLE_CURRENT = 'Current';
-const RUN_TITLE_BASELINE = 'Baseline';
-
-const getAddRunLabel = (items) => (run, index, runs) => {
+const getAddRunLabel = (run, index, runs) => {
   const internalBuildNumber = get(run, 'meta.internalBuildNumber', runs.length - index);
   const name = `Job #${internalBuildNumber}`;
-
-  const label = (
-    <div className={css.tableHeaderRun}>
-      <JobName
-        title={index === 0 ? RUN_TITLE_CURRENT : RUN_TITLE_BASELINE}
-        internalBuildNumber={internalBuildNumber}
-      />
-      <RunLabelSum
-        className={css.tableHeaderRunMetric}
-        runIndex={index}
-        runCount={runs.length}
-        rows={items}
-      />
-    </div>
-  );
 
   return {
     ...run,
     name,
-    label,
   };
 };
 
@@ -55,7 +34,8 @@ export const BundlePackages = (props) => {
     updateSort,
   } = props;
 
-  const labeledRuns = runs.map(getAddRunLabel(items));
+  const labeledRuns = runs.map(getAddRunLabel);
+
   const emptyMessage = (
     <EmptySet
       resources="packages"
@@ -97,6 +77,7 @@ export const BundlePackages = (props) => {
           runs={labeledRuns}
           items={items}
           emptyMessage={emptyMessage}
+          showHeaderSum
         />
       </main>
     </section>
