+++ packages/ui/src/components/bundle-packages/bundle-packages.module.css
@@ -23,15 +23,6 @@
   margin-left: var(--space-small);
 }
 
-.tableHeaderRun {
-  margin-top: calc(0px - var(--space-xxsmall));
-  margin-bottom: calc(0px - var(--space-xxsmall));
-}
-
-.tableHeaderRunMetric {
-  margin-top: var(--space-xxxsmall);
-}
-
 /* Tooltip filename */
 .tooltipFilename {
   text-align: left;
